refactor presign url logic out of seed_data file into models file. Now called on request
