+++ Forward-client/.env
@@ -1,3 +1,3 @@
 DEBUG=false
-VITE_BACKEND_URL=http://localhost:8000
+VITE_BACKEND_URL=http://backend:8000
 # VITE_BACKEND_URL=https://backend-latest-ma6c.onrender.com
